+++ Private/Classes/LoggerService.ps1
@@ -44,6 +44,7 @@ class LoggerService {
                     'console' { Invoke-SMConsoleLogger $Component $Message $Severity $IndentLevel }
                     'file'    { Invoke-FileSink -Options $sink.Options -Component $Component -Message $Message -Severity $Severity -IndentLevel $IndentLevel }
                     'web'     { Invoke-WebSink  -Options $sink.Options -Component $Component -Message $Message -Severity $Severity -IndentLevel $IndentLevel }
+                    'serilog' { Invoke-SerilogSink -Options $sink.Options -Component $Component -Message $Message -Severity $Severity -IndentLevel $IndentLevel }
                 }
             }
         }