+++ Public/Register-LoggerSink.ps1
@@ -1,7 +1,7 @@
 function Register-LoggerSink {
     [CmdletBinding(DefaultParameterSetName='Console')]
     param(
-        [Parameter(Mandatory)][ValidateSet('Console','File','Web')][string]$Type,
+        [Parameter(Mandatory)][ValidateSet('Console','File','Web','Serilog')][string]$Type,
 
         # Console
         [Parameter(ParameterSetName='Console')][ValidateSet('Default')][string]$Format = 'Default',
@@ -14,13 +14,16 @@ function Register-LoggerSink {
         [Parameter(ParameterSetName='File')][ValidateRange(1,100)][int]$MaxRolls = 3,
         [Parameter(ParameterSetName='File')][ValidateSet('UTF8','UTF8BOM','ASCII','Unicode','UTF7','UTF32','Default','OEM')][string]$Encoding = 'UTF8BOM',
         [Parameter(ParameterSetName='File')]
-        [Parameter(ParameterSetName='Web')]
+        [Parameter(ParameterSetName='Http')]
         [ValidateSet('Warn','Continue','Throw')][string]$OnError = 'Warn',
 
-        # Web
-        [Parameter(Mandatory, ParameterSetName='Web')][string]$Url,
-        [Parameter(ParameterSetName='Web')][string]$APIKey,
-        [Parameter(ParameterSetName='Web')][hashtable]$Headers
+        # HTTP sinks
+        [Parameter(Mandatory, ParameterSetName='Http')]
+        [string]$Url,
+        [Parameter(ParameterSetName='Http')]
+        [string]$APIKey,
+        [Parameter(ParameterSetName='Http')]
+        [hashtable]$Headers
     )
 
     $svc = [LoggerService]::GetInstance()
@@ -33,6 +36,9 @@ function Register-LoggerSink {
         }
         'Web' {
             $svc.RegisterSink('Web', @{ Url = $Url; APIKey = $APIKey; Headers = $Headers; OnError = $OnError })
+        }
+        'Serilog' {
+            $svc.RegisterSink('Serilog', @{ Url = $Url; APIKey = $APIKey; Headers = $Headers; OnError = $OnError })
         }
     }
 }